+++ source/MaterialXFormat/XmlIo.cpp
@@ -28,8 +28,6 @@ const string XINCLUDE_URL = "http://www.w3.org/2001/XInclude";
 
 void elementFromXml(const xml_node& xmlNode, ElementPtr elem, const XmlReadOptions* readOptions)
 {
-    bool skipConflictingElements = readOptions && readOptions->skipConflictingElements;
-
     // Store attributes in element.
     for (const xml_attribute& xmlAttr : xmlNode.attributes())
     {
@@ -55,7 +53,7 @@ void elementFromXml(const xml_node& xmlNode, ElementPtr elem, const XmlReadOptio
 
         // Check for duplicate elements.
         ConstElementPtr previous = elem->getChild(name);
-        if (previous && skipConflictingElements)
+        if (previous)
         {
             continue;
         }
@@ -63,12 +61,6 @@ void elementFromXml(const xml_node& xmlNode, ElementPtr elem, const XmlReadOptio
         // Create the new element.
         ElementPtr child = elem->addChildOfCategory(category, name, !previous);
         elementFromXml(xmlChild, child, readOptions);
-
-        // Check for conflicting elements.
-        if (previous && *previous != *child)
-        {
-            throw Exception("Duplicate element with conflicting content: " + name);
-        }
     }
 }
 
@@ -208,7 +200,7 @@ void processXIncludes(DocumentPtr doc, xml_node& xmlNode, const FileSearchPath& 
                 readXIncludeFunction(library, filename, includeSearchPath, &xiReadOptions);
 
                 // Import the library document.
-                doc->importLibrary(library, readOptions);
+                doc->importLibrary(library);
             }
 
             // Remove include directive.