+++ source/MaterialXTest/MaterialXRender/RenderUtil.cpp
@@ -214,9 +214,6 @@ bool ShaderRenderTester::validate(const mx::FilePathVec& testRootPaths, const mx
 
     mx::StringSet usedImpls;
 
-    mx::CopyOptions copyOptions;
-    copyOptions.skipConflictingElements = true;
-
     const std::string MTLX_EXTENSION("mtlx");
     for (const auto& dir : dirs)
     {
@@ -257,7 +254,7 @@ bool ShaderRenderTester::validate(const mx::FilePathVec& testRootPaths, const mx
                 WARN("Failed to load in file: " + filename.asString() + "See: " + docValidLogFilename + " for details.");
             }
 
-            doc->importLibrary(dependLib, &copyOptions);
+            doc->importLibrary(dependLib);
             ioTimer.endTimer();
 
             validateTimer.startTimer();